Refactor Iex Connection to separate Class- Test Connector- Refactor Processor, ReTest
