+++ test/workflows/iex_fetch_daily_data_test_processor.rb
@@ -3,33 +3,23 @@ require 'minitest/autorun'
 
 class IexFetchDailyDataProcessorTest < ActiveSupport::TestCase
 
-  test 'valid quote returns' do
-    VCR.use_cassette('iex_fetch_daily_data_valid_quote') do
-      @iex = IexFetchDailyDataProcessor.new
-      @quote = @iex.quote('AAPL')
+  setup do
+    VCR.use_cassette('iex_connector_test') do
+      @iex_connector = IexConnector.new
+      @quote = @iex_connector.get_quote('AAPL')
+      @iex_fetch_daily_data = IexFetchDailyDataProcessor.new
     end
-    assert_equal 'AAPL', @quote.symbol
-  end
-
-  test 'invalid quote' do
-    VCR.use_cassette('iex_fetch_daily_data_invalid_quote') do
-      @iex = IexFetchDailyDataProcessor.new
-      @quote = @iex.quote('ZZZ')
-    end
-    assert_equal nil, @quote
-    assert_equal ['ZZZ'], @iex.symbols_not_found
   end
 
   test '#parse_quote_body returns formatted object' do
     VCR.use_cassette('iex_daily_date_parse_body') do
-      @iex = IexFetchDailyDataProcessor.new
-      @quote = @iex.quote('AAPl')
-      @body = @iex.parse_quote_body(@quote)
+      @iex_fetch_daily_data.quote_body = @quote
+      @body = @iex_fetch_daily_data.parse_quote_body
     end
 
     expected = {:market_close_date=>"07-01-2019",
-      :open=>144.5, :high=>148.55, :low=>143.8,
-      :close=>148.26, :net_change=>6.07, :volume=>57288976}
+      :open=>148.7, :high=>148.83, :low=>145.9, :close=>148.26,
+      :net_change=>-1.035, :volume=>31077781}
     assert_equal expected, @body
   end
